refactor: changes all ui components
