--- src/components/Form/Form.hbs
@@ -1,6 +1,0 @@
-<form class="{{settings.root.className}}" {{#if settings.root.action}}action="{{settings.root.action}}"{{/if}}>
-    {{#if settings.title.text}}<h3 class="{{settings.title.className}}">{{settings.title.text}}</h3>{{/if}}
-    {{#each settings.fields}}
-        {{{this}}}
-    {{/each}}
-</form>
