--- src/components/index.js
@@ -1,1 +1,0 @@
-export { default as Form } from './Form';
