+++ src/ui/Button/Button.hbs
@@ -1,1 +1,1 @@
-<button class="{{className}}">{{text}}</button>
+<button class="button">{{button.text}}</button>
