+++ src/ui/Button/Button.scss
@@ -1,4 +1,4 @@
-.root {
+.button {
     border: none;
     border-radius: 5px;
     background-color: #3369F3;
