+++ src/ui/Form/Form.hbs
@@ -1,0 +1,4 @@
+<form class="form">
+    {{#if title}}<h3 class="form__title">{{title}}</h3>{{/if}}
+    {{> @partial-block}}
+</form>
