+++ src/ui/Form/Form.scss
@@ -1,4 +1,4 @@
-.root {
+.form {
     display: flex;
     flex-direction: column;
     gap: 16px;
@@ -8,7 +8,7 @@
     padding: 30px;
     box-shadow: 0px 0px 6px #00000024;
 
-    .title {
+    .form__title {
         font-size: 20px;
         font-weight: 500;
         font-family: "Inter", sans-serif;
