+++ src/ui/Form/index.js
@@ -1,0 +1,2 @@
+import './Form.scss';
+export { default as Form } from './Form.hbs?raw';
