+++ src/ui/Input/Input.hbs
@@ -1,8 +1,9 @@
-<div class="{{settings.root.className}}">
-    <label for="{{setting.label.for}}" class="{{settings.label.className}}">{{settings.label.label}}</label>
-    <input class="{{settings.input.className}}" 
-    {{#if settings.input.placeholder}} placeholder="{{settings.input.placeholder}}" {{/if}} 
-    {{#if settings.input.id}} id="{{settings.input.id}}" {{/if}} 
-    {{#if settings.input.name}} name="{{settings.input.name}}" {{/if}}
-    {{#if settings.input.type}} type="{{settings.input.type}}" {{/if}}>
+<div class="input">
+    {{#if input.label}}<label for="{{input.name}}" class="input__label">{{input.label}}</label>{{/if}}
+    <input 
+    class="input__input" 
+    placeholder="{{input.placeholder}}"
+    value="{{#if input.value}}{{input.value}}{{/if}}"
+    name="{{input.name}}"
+    type="{{input.type}}">
 </div>
