+++ src/ui/Input/index.js
@@ -1,1 +1,2 @@
-export { default } from './Input';
+import './Input.scss';
+export { default as Input } from './Input.hbs?raw';
