+++ src/ui/index.js
@@ -1,2 +1,3 @@
-export { default as Button } from './Button';
-export { default as Input } from './Input';
+export * from './Button';
+export * from './Form';
+export * from './Input';
